+++ RELEASE_NOTES.md
@@ -1,7 +1,13 @@
 Release Notes
 =============
+
 ## 1.6.5
+* Azure Firewall: Bug fix for link_to_vhub and added depends_on to builder
+* Functions: Add support for keyvault reference user identity
 * VirtualHubs/hubRouteTables : Add support for labels
+* Virtual Machine: Add option to static IP allocation
+* Web App: Add support for keyvault reference user identity
+
 ## 1.6.4
 * DNS Zone: Added SOA and SRV record support
 * Azure Firewall: Added support for Azure Firewalls